[kernel] Build everything with -Wconversion [part 20]
Changes to be committed:
	modified:   include/tilck/common/color_defs.h
	modified:   include/tilck/common/debug/termios_debug.c.h
	modified:   kernel/char/tty/tty_ioctl.c
	modified:   kernel/char/tty/tty_output.c

+++ include/tilck/common/color_defs.h
@@ -28,7 +28,7 @@ enum vga_color {
    COLOR_BRIGHT_WHITE = 15,
 };
 
-#define make_color(fg, bg) ((fg) | (bg) << 4)
+#define make_color(fg, bg) ((u8)(((fg) | (bg) << 4)))
 #define get_color_fg(color) ((color) & 0xF)
 #define get_color_bg(color) (((color) >> 4) & 0xF)
